+++ src/main/java/com/example/demo/domain/ReviewImpl.java
@@ -15,6 +15,7 @@ import javax.persistence.Table;
 
 import org.thymeleaf.util.StringUtils;
 
+import com.example.demo.domain.cafe.CafeImpl;
 import com.example.demo.exception.CafegoryException;
 
 import lombok.AccessLevel;
