+++ src/main/java/com/example/demo/domain/cafe/Cafe.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain;
+package com.example.demo.domain.cafe;
 
 public interface Cafe {
 	String showFullAddress();
