+++ src/main/java/com/example/demo/domain/cafe/CafeSearchCondition.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain;
+package com.example.demo.domain.cafe;
 
 import static com.example.demo.exception.ExceptionType.*;
