+++ src/main/java/com/example/demo/domain/cafe/Menu.java
@@ -1,5 +1,6 @@
-package com.example.demo.domain;
+package com.example.demo.domain.cafe;
 
+import javax.persistence.Column;
 import javax.persistence.ConstraintMode;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -18,16 +19,19 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @Builder
+@Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "sns_detail")
-public class SnsDetail {
+@Table(name = "menu")
+public class Menu {
+
 	@Id
 	@GeneratedValue
+	@Column(name = "menu_id")
 	private Long id;
+
 	private String name;
-	private String url;
+	private int price;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "cafe_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
