+++ src/main/java/com/example/demo/mapper/BusinessHourMapper.java
@@ -3,7 +3,7 @@ package com.example.demo.mapper;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.example.demo.domain.BusinessHour;
+import com.example.demo.domain.cafe.BusinessHour;
 import com.example.demo.dto.BusinessHourResponse;
 
 public class BusinessHourMapper {
