+++ src/main/java/com/example/demo/mapper/SnsDetailMapper.java
@@ -3,7 +3,7 @@ package com.example.demo.mapper;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.example.demo.domain.SnsDetail;
+import com.example.demo.domain.cafe.SnsDetail;
 import com.example.demo.dto.SnsResponse;
 
 public class SnsDetailMapper {
