+++ src/main/java/com/example/demo/repository/cafe/CafeRepository.java
@@ -2,7 +2,7 @@ package com.example.demo.repository.cafe;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.demo.domain.CafeImpl;
+import com.example.demo.domain.cafe.CafeImpl;
 
 public interface CafeRepository extends JpaRepository<CafeImpl, Long>, CafeRepositoryCustom {
 }
