+++ src/test/java/com/example/demo/domain/cafe/BusinessHourOpenCheckerTest.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain;
+package com.example.demo.domain.cafe;
 
 import static org.assertj.core.api.Assertions.*;
