+++ src/test/java/com/example/demo/repository/cafe/CafeRepositorySearchMethodTest.java
@@ -22,15 +22,15 @@ import org.springframework.data.domain.Page;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.demo.config.QueryDslConfig;
-import com.example.demo.domain.Address;
-import com.example.demo.domain.BusinessHour;
-import com.example.demo.domain.CafeImpl;
-import com.example.demo.domain.CafeSearchCondition;
-import com.example.demo.domain.MaxAllowableStay;
 import com.example.demo.domain.MemberImpl;
-import com.example.demo.domain.Menu;
 import com.example.demo.domain.ReviewImpl;
-import com.example.demo.domain.SnsDetail;
+import com.example.demo.domain.cafe.Address;
+import com.example.demo.domain.cafe.BusinessHour;
+import com.example.demo.domain.cafe.CafeImpl;
+import com.example.demo.domain.cafe.CafeSearchCondition;
+import com.example.demo.domain.cafe.MaxAllowableStay;
+import com.example.demo.domain.cafe.Menu;
+import com.example.demo.domain.cafe.SnsDetail;
 import com.example.demo.util.PageRequestCustom;
 
 // @SpringBootTest
