+++ src/test/java/com/example/demo/service/ReviewServiceTest.java
@@ -12,10 +12,10 @@ import org.springframework.context.annotation.Import;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.demo.config.TestConfig;
-import com.example.demo.domain.CafeImpl;
 import com.example.demo.domain.MemberImpl;
 import com.example.demo.domain.ReviewImpl;
 import com.example.demo.domain.ThumbnailImage;
+import com.example.demo.domain.cafe.CafeImpl;
 import com.example.demo.dto.ReviewSaveRequest;
 import com.example.demo.dto.ReviewUpdateRequest;
 import com.example.demo.exception.CafegoryException;
