+++ src/test/java/com/example/demo/service/StudyOnceCommentQueryServiceImplTest.java
@@ -14,11 +14,11 @@ import org.springframework.context.annotation.Import;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.demo.config.TestConfig;
-import com.example.demo.domain.CafeImpl;
 import com.example.demo.domain.MemberImpl;
 import com.example.demo.domain.StudyOnceComment;
 import com.example.demo.domain.StudyOnceImpl;
 import com.example.demo.domain.ThumbnailImage;
+import com.example.demo.domain.cafe.CafeImpl;
 import com.example.demo.dto.MemberResponse;
 import com.example.demo.dto.StudyOnceCommentSearchResponse;
 import com.example.demo.dto.StudyOnceCommentsSearchResponse;
