+++ src/test/java/com/example/demo/service/StudyOnceServiceImplTest.java
@@ -23,14 +23,14 @@ import org.springframework.context.annotation.Import;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.demo.config.TestConfig;
-import com.example.demo.domain.Address;
 import com.example.demo.domain.Attendance;
-import com.example.demo.domain.CafeImpl;
 import com.example.demo.domain.MemberImpl;
 import com.example.demo.domain.StudyMember;
 import com.example.demo.domain.StudyMemberId;
 import com.example.demo.domain.StudyOnceImpl;
 import com.example.demo.domain.ThumbnailImage;
+import com.example.demo.domain.cafe.Address;
+import com.example.demo.domain.cafe.CafeImpl;
 import com.example.demo.dto.MemberResponse;
 import com.example.demo.dto.PagedResponse;
 import com.example.demo.dto.StudyMemberStateRequest;
